เปลี่ยน Location จาก int เป็น double

+++ domain/src/main/java/th/or/nectec/tanrabad/domain/geographic/FilterBoundaryCalculate.java
@@ -22,7 +22,7 @@ import th.or.nectec.tanrabad.entity.Location;
 public class FilterBoundaryCalculate implements FilterBoundaryCalculator {
 
     @Override
-    public Location getMinLocation(Location currentLocation, int distanceInKm) {
+    public Location getMinLocation(Location currentLocation, double distanceInKm) {
         double longitudeDegreeFromKm  = distanceInKm/111.699;
         double latitudeDegreeFromKm = distanceInKm/110.567;
         double minimumLongitude = currentLocation.getLongitude() - longitudeDegreeFromKm;
@@ -32,7 +32,7 @@ public class FilterBoundaryCalculate implements FilterBoundaryCalculator {
     }
 
     @Override
-    public Location getMaxLocation(Location currentLocation, int distanceInKm) {
+    public Location getMaxLocation(Location currentLocation, double distanceInKm) {
         double longitudeDegreeFromKm  = distanceInKm/111.699;
         double latitudeDegreeFromKm = distanceInKm/110.567;
         double maximumLongitude = currentLocation.getLongitude() + longitudeDegreeFromKm;
